r2086/sofa-dev : FIX: warnings on Windows
 //partial sync (9/11 changes)//

+++ modules/sofa/component/constraint/SlidingConstraint.inl
@@ -165,7 +165,7 @@ void SlidingConstraint<DataTypes>::getConstraintId(long* id, unsigned int &offse
 {
     if (!yetIntegrated)
     {
-        id[offset++] = -cid;
+        id[offset++] = -(int)cid;
 
         yetIntegrated =  true;
     }
